+++ model_compression_toolkit/gptq/common/gptq_framework_implementation.py
@@ -1,0 +1,32 @@
+# Copyright 2023 Sony Semiconductor Israel, Inc. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+
+from abc import abstractmethod
+
+from model_compression_toolkit.core.common.framework_implementation import FrameworkImplementation
+
+
+class GPTQFrameworkImplemantation(FrameworkImplementation):
+    """
+    Class to implement framework related methods that are used in GPTQ
+    """
+
+    @abstractmethod
+    def get_gptq_trainer_obj(self):
+        """
+        Returns: GPTQTrainer object
+        """
+        raise NotImplemented(f'{self.__class__.__name__} have to implement the '
+                             f'framework\'s get_gptq_trainer method.')  # pragma: no cover
