+++ tests/quantizers_infrastructure_tests/trainable_infrastructure_tests/keras/test_keras_trainable_infra_runner.py
@@ -15,7 +15,7 @@
 import unittest
 import tensorflow as tf
 
-from model_compression_toolkit import TrainingMethod
+from model_compression_toolkit.qat import TrainingMethod
 from model_compression_toolkit.target_platform_capabilities.target_platform import QuantizationMethod
 from model_compression_toolkit.qat.keras.quantizer.ste_rounding.symmetric_ste import STEWeightQATQuantizer, \
     STEActivationQATQuantizer
