+++ tests/quantizers_infrastructure_tests/trainable_infrastructure_tests/pytorch/test_pytorch_trainable_infra_runner.py
@@ -16,7 +16,7 @@
 
 import unittest
 
-from model_compression_toolkit import TrainingMethod
+from model_compression_toolkit.qat import TrainingMethod
 from model_compression_toolkit.target_platform_capabilities.target_platform import QuantizationMethod
 from model_compression_toolkit.qat.pytorch.quantizer.ste_rounding.symmetric_ste import STEWeightQATQuantizer, \
     STEActivationQATQuantizer
